fix typo

+++ tests/managers/test_managers_inherited.py
@@ -65,7 +65,8 @@ async def create_test_database():
     async with database:
         await models.create_all()
         yield
-        await models.drop_all()
+        if not database.drop:
+            await models.drop_all()
 
 
 @pytest.fixture(autouse=True, scope="function")
